removed dependency of lookback years on result years
Removed dependency of lookbacks on years for which we are computing the
results.  In short, it is now possible to compute results for just one
year, or two years, etc.  and still have access to lookbacks going back
two years, without having to compute results for these lookback years as
well.